Move components to pages directory
