+++ app/app.routing.ts
@@ -2,10 +2,10 @@ import { NgModule } from '@angular/core';
 import { NativeScriptRouterModule } from 'nativescript-angular/router';
 import { Routes } from '@angular/router';
 
-import { TaskListComponent } from './components/task-list/task-list.component';
-import { TaskFormComponent } from './components/task-form/task-form.component';
-import { TaskDetailComponent } from '~/components/task-detail/task-detail.component';
-import { LoginMainComponent } from '~/components/login-main/login-main.component';
+import { TaskListComponent } from './pages/tasks/task-list/task-list.component';
+import { TaskFormComponent } from './pages/tasks/task-form/task-form.component';
+import { TaskDetailComponent } from '~/pages/tasks/task-detail/task-detail.component';
+import { LoginMainComponent } from '~/pages/account/login-main/login-main.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/task-list', pathMatch: 'full' },
@@ -13,7 +13,7 @@ const routes: Routes = [
   { path: 'task-form', component: TaskFormComponent },
   { path: 'task-detail/:id', component: TaskDetailComponent },
   { path: 'task-edit/:id', component: TaskFormComponent },
-  { path: 'login-main', component: LoginMainComponent}
+  { path: 'login-main', component: LoginMainComponent }
 ];
 
 @NgModule({
