+++ app/pages/tasks/tasks.module.ts
@@ -1,7 +1,7 @@
 import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { NativeScriptCommonModule } from 'nativescript-angular/common';
 
-import { TaskListComponent } from './task-list.component';
+import { TaskListComponent } from './task-list/task-list.component';
 
 @NgModule({
   imports: [NativeScriptCommonModule],
